requirements.txt
asgiref==3.7.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
Django==4.2
google-trans-new==1.1.9
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
djangorestframework==3.14.0
drf-spectacular==0.26.2
celery==5.3.1
redis==4.5.5
django-celery-results